backend/chatboot/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
fastapi>=0.116.0
uvicorn>=0.35.0
pandas>=2.2.0
scikit-learn>=1.6.0
matplotlib>=3.9.0
seaborn>=0.13.0
python-multipart>=0.0.9

# Embedding RAG dependencies
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4

